get post by id

+++ client/src/actions/post.js
@@ -6,6 +6,8 @@ import {
   UPDATE_LIKES,
   DELETE_POST,
   ADD_POST,
+  CONSOLE_LOG_ERROR,
+  GET_POST,
 } from "./types";
 
 const URL = "http://localhost:5000";
@@ -19,10 +21,21 @@ export const getPosts = () => async (dispatch) => {
       payload: res.data,
     });
   } catch (err) {
-    dispatch({
-      type: POST_ERROR,
-      payload: err.response,
-    });
+    if (err.response) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: err.response.statusText,
+          status: err.response.status,
+          data: err.response.data,
+        },
+      });
+    } else {
+      console.log(err);
+      dispatch({
+        type: CONSOLE_LOG_ERROR,
+      });
+    }
   }
 };
 
@@ -35,10 +48,21 @@ export const addLike = (postId) => async (dispatch) => {
       payload: { postId, likes: res.data },
     });
   } catch (err) {
-    dispatch({
-      type: POST_ERROR,
-      payload: err.response,
-    });
+    if (err.response) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: err.response.statusText,
+          status: err.response.status,
+          data: err.response.data,
+        },
+      });
+    } else {
+      console.log(err);
+      dispatch({
+        type: CONSOLE_LOG_ERROR,
+      });
+    }
   }
 };
 
@@ -51,28 +75,52 @@ export const removeLike = (postId) => async (dispatch) => {
       payload: { postId, likes: res.data },
     });
   } catch (err) {
-    dispatch({
-      type: POST_ERROR,
-      payload: err.response,
-    });
+    if (err.response) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: err.response.statusText,
+          status: err.response.status,
+          data: err.response.data,
+        },
+      });
+    } else {
+      console.log(err);
+      dispatch({
+        type: CONSOLE_LOG_ERROR,
+      });
+    }
   }
 };
 
 //delete post
 export const deletePost = (postId) => async (dispatch) => {
-  try {
-    await axios.delete(`${URL}/api/posts/${postId}`);
-    dispatch({
-      type: DELETE_POST,
-      payload: postId,
-    });
+  if (window.confirm("Are you sure? You want to delete this post.")) {
+    try {
+      await axios.delete(`${URL}/api/posts/${postId}`);
+      dispatch({
+        type: DELETE_POST,
+        payload: postId,
+      });
 
-    dispatch(setAlert("Post Remove", "success"));
-  } catch (err) {
-    dispatch({
-      type: POST_ERROR,
-      payload: err.response,
-    });
+      dispatch(setAlert("Post Remove", "success"));
+    } catch (err) {
+      if (err.response) {
+        dispatch({
+          type: POST_ERROR,
+          payload: {
+            msg: err.response.statusText,
+            status: err.response.status,
+            data: err.response.data,
+          },
+        });
+      } else {
+        console.log(err);
+        dispatch({
+          type: CONSOLE_LOG_ERROR,
+        });
+      }
+    }
   }
 };
 
@@ -93,9 +141,47 @@ export const addPost = (formData) => async (dispatch) => {
 
     dispatch(setAlert("Post Created", "success"));
   } catch (err) {
+    if (err.response) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: err.response.statusText,
+          status: err.response.status,
+          data: err.response.data,
+        },
+      });
+    } else {
+      console.log(err);
+      dispatch({
+        type: CONSOLE_LOG_ERROR,
+      });
+    }
+  }
+};
+
+//Get a posts
+export const getPost = (postId) => async (dispatch) => {
+  try {
+    const res = await axios.get(`${URL}/api/posts/${postId}`);
     dispatch({
-      type: POST_ERROR,
-      payload: err.response,
+      type: GET_POST,
+      payload: res.data,
     });
+  } catch (err) {
+    if (err.response) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: err.response.statusText,
+          status: err.response.status,
+          data: err.response.data,
+        },
+      });
+    } else {
+      console.log(err);
+      dispatch({
+        type: CONSOLE_LOG_ERROR,
+      });
+    }
   }
 };
